Added recording of best tigers per epoch & death walls
Creatures now die if they leave the game area.

A best tiger list is maintained on creature death - if it was a tiger and it was good enough for the genepool, add it to the list. The list is cleared at the start of each epoch, so only new breeders from each epoch will be saved. Only the first 100 bases of DNA are recorded to save filespace.